+++ common/nymsphinx/src/lib.rs
@@ -1,0 +1,9 @@
+pub mod addressing;
+pub mod chunking;
+
+// Future consideration: currently in a lot of places, the payloads have randomised content
+// which is not a perfect testing strategy as it might not detect some edge cases I never would
+// have assumed could be possible. A better approach would be to research some Fuzz testing
+// library like: https://github.com/rust-fuzz/afl.rs and use that instead for the inputs.
+
+// perhaps it might be useful down the line for interaction testing between client,mixes,etc?