+++ mixnode/src/node/mod.rs
@@ -2,11 +2,13 @@ use crate::config::persistence::pathfinder::MixNodePathfinder;
 use crate::config::Config;
 use crate::node::packet_processing::PacketProcessor;
 use crypto::encryption;
+use directory_client::presence::Topology;
 use futures::channel::mpsc;
 use log::*;
 use pemstore::pemstore::PemStore;
 use std::net::SocketAddr;
 use tokio::runtime::Runtime;
+use topology::NymTopology;
 
 mod listener;
 mod metrics;
@@ -62,6 +64,7 @@ impl MixNode {
             self.config.get_metrics_directory_server(),
             self.sphinx_keypair.public_key().to_base58_string(),
             self.config.get_metrics_sending_delay(),
+            self.config.get_metrics_running_stats_logging_delay(),
         )
         .start(self.runtime.handle())
     }
@@ -92,16 +95,38 @@ impl MixNode {
                 packet_forwarding::PacketForwarder::new(
                     self.config.get_packet_forwarding_initial_backoff(),
                     self.config.get_packet_forwarding_maximum_backoff(),
+                    self.config.get_initial_connection_timeout(),
                 )
             })
             .start(self.runtime.handle())
     }
 
+    fn check_if_same_ip_node_exists(&self) -> Option<String> {
+        // TODO: once we change to graph topology this here will need to be updated!
+        let topology = Topology::new(self.config.get_presence_directory_server());
+        let existing_mixes_presence = topology.mix_nodes;
+        existing_mixes_presence
+            .iter()
+            .find(|node| node.host == self.config.get_announce_address())
+            .map(|node| node.pub_key.clone())
+    }
+
     pub fn run(&mut self) {
+        info!("Starting nym mixnode");
+
+        if let Some(duplicate_node_key) = self.check_if_same_ip_node_exists() {
+            error!(
+                "Our announce-host is identical to one of existing nodes! (its key is {:?}",
+                duplicate_node_key
+            );
+            return;
+        }
         let forwarding_channel = self.start_packet_forwarder();
         let metrics_reporter = self.start_metrics_reporter();
         self.start_socket_listener(metrics_reporter, forwarding_channel);
         self.start_presence_notifier();
+
+        info!("Finished nym mixnode startup procedure - it should now be able to receive mix traffic!");
 
         if let Err(e) = self.runtime.block_on(tokio::signal::ctrl_c()) {
             error!(