+++ nym-client/src/config/mod.rs
@@ -20,6 +20,8 @@ const DEFAULT_TOPOLOGY_REFRESH_RATE: u64 = 10_000;
 const DEFAULT_TOPOLOGY_RESOLUTION_TIMEOUT: u64 = 5_000;
 const DEFAULT_PACKET_FORWARDING_INITIAL_BACKOFF: u64 = 10_000; // 10s
 const DEFAULT_PACKET_FORWARDING_MAXIMUM_BACKOFF: u64 = 300_000; // 5min
+const DEFAULT_INITIAL_CONNECTION_TIMEOUT: u64 = 1_500; // 1.5s
+const DEFAULT_HEALTHCHECK_CONNECTION_TIMEOUT: u64 = DEFAULT_INITIAL_CONNECTION_TIMEOUT;
 
 const DEFAULT_NUMBER_OF_HEALTHCHECK_TEST_PACKETS: u64 = 2;
 const DEFAULT_NODE_SCORE_THRESHOLD: f64 = 0.0;
@@ -213,6 +215,14 @@ impl Config {
     pub fn get_packet_forwarding_maximum_backoff(&self) -> time::Duration {
         time::Duration::from_millis(self.debug.packet_forwarding_maximum_backoff)
     }
+
+    pub fn get_initial_connection_timeout(&self) -> time::Duration {
+        time::Duration::from_millis(self.debug.initial_connection_timeout)
+    }
+
+    pub fn get_healthcheck_connection_timeout(&self) -> time::Duration {
+        time::Duration::from_millis(self.debug.healthcheck_connection_timeout)
+    }
 }
 
 fn de_option_string<'de, D>(deserializer: D) -> Result<Option<String>, D::Error>
@@ -317,7 +327,7 @@ impl Default for Logging {
 }
 
 #[derive(Debug, Deserialize, PartialEq, Serialize)]
-#[serde(deny_unknown_fields)]
+#[serde(default, deny_unknown_fields)]
 pub struct Debug {
     /// The parameter of Poisson distribution determining how long, on average,
     /// sent packet is going to be delayed at any given mix node.
@@ -376,6 +386,15 @@ pub struct Debug {
     /// forwarding sphinx packets.
     /// The provided value is interpreted as milliseconds.
     packet_forwarding_maximum_backoff: u64,
+
+    /// Timeout for establishing initial connection when trying to forward a sphinx packet.
+    /// The provider value is interpreted as milliseconds.
+    initial_connection_timeout: u64,
+
+    /// Timeout for establishing initial connection when trying to forward a sphinx packet
+    /// during healthcheck.
+    /// The provider value is interpreted as milliseconds.
+    healthcheck_connection_timeout: u64,
 }
 
 impl Default for Debug {
@@ -392,6 +411,8 @@ impl Default for Debug {
             node_score_threshold: DEFAULT_NODE_SCORE_THRESHOLD,
             packet_forwarding_initial_backoff: DEFAULT_PACKET_FORWARDING_INITIAL_BACKOFF,
             packet_forwarding_maximum_backoff: DEFAULT_PACKET_FORWARDING_MAXIMUM_BACKOFF,
+            initial_connection_timeout: DEFAULT_INITIAL_CONNECTION_TIMEOUT,
+            healthcheck_connection_timeout: DEFAULT_HEALTHCHECK_CONNECTION_TIMEOUT,
         }
     }
 }