+++ validator/src/config/template.rs
@@ -43,6 +43,10 @@ run_delay = {{ mix_mining.run_delay }}
 # The provided value is interpreted as milliseconds.
 resolution_timeout = {{ mix_mining.resolution_timeout }}
 
+# Timeout for trying to establish connection to node endpoints.
+# The provided value is interpreted as milliseconds.
+connection_timeout =  {{ mix_mining.connection_timeout }}
+    
 # How many packets should be sent through each path during the mix-mining procedure.
 number_of_test_packets = {{ mix_mining.number_of_test_packets }}
 
@@ -58,19 +62,6 @@ number_of_test_packets = {{ mix_mining.number_of_test_packets }}
 [logging]
 
 # TODO
-
-
-##### debug configuration options #####
-# The following options should not be modified unless you know EXACTLY what you are doing
-# as if set incorrectly, they may impact your anonymity.
-
-[debug]
-
-# Directory server to which the server will be reporting their presence data.
-presence_directory_server = '{{ debug.presence_directory_server }}'
-
-# Delay between each subsequent presence data being sent.
-presence_sending_delay = {{ debug.presence_sending_delay }}
 
 "#
 }